// types/pointOutcomes.ts
export interface PointOutcomes {
    pointsWonByForehand: number;
    pointsWonByBackhand: number;
    pointsWonByForehandVolley: number;
    pointsWonByBackhandVolley: number;
    pointsWonByOverhead: number;
    pointsWonByForehandWinner: number;
    pointsWonByBackhandWinner: number;
    pointsWonByForehandVolleyWinner: number;
    pointsWonByBackhandVolleyWinner: number;
    pointsWonByOverheadWinner: number;
    pointsWonOnServe: number;
    pointsWonOnReturn: number;
    pointsLostByForehand: number;
    pointsLostByBackhand: number;
    pointsLostByForehandVolley: number;
    pointsLostByBackhandVolley: number;
    pointsLostByOverhead: number;
    pointsLostByForehandUnforcedError: number;
    pointsLostByBackhandUnforcedError: number;
    pointsLostByForehandVolleyUnforcedError: number;
    pointsLostByBackhandVolleyUnforcedError: number;
    pointsLostByOverheadUnforcedError: number;
    pointsLostOnServe: number;
    pointsLostOnReturn: number;
    missedFirstServeReturns: number;
    missedSecondServeReturns: number;
    missedFirstServes: number;
    doubleFaults: number;
    acesOnFirstServe: number;
    acesOnSecondServe: number;
    madeFirstServes: number;
    madeSecondServes: number;
    notReturnedFirstServes: number;
    notReturnedSecondServes: number;
    notReturnedServesDeuceSide: number;
    notReturnedServesAdSide: number;
    notReturnedServesToAlley: number;
    notReturnedServesToBody: number;
    notReturnedServesToCenter: number;
}